- Changement des réglages de l'envoi de mail suite à changement de smtp - souci sur parameters.yml -> impossible de récupérer la bonne valeur de nb_per_page

+++ src/PB/TransportBundle/Entity/Transport.php
@@ -137,7 +137,7 @@ class Transport
     /**
      * @var int
      *
-     *@ORM\Column(name="mplancher", type="integer", nullable=true)
+     *@ORM\Column(name="mplancher", type="decimal", nullable=true)
      * @ORM\JoinColumn(nullable=true)
      */
     private $mplancher; 
@@ -195,6 +195,8 @@ class Transport
      * @return int
      */
     
+
+
     public function __construct()
     {
         $now = new \DateTime();
